chore(reorg): move docs catalog/provider into content ownership
