+++ src/content/index.ts
@@ -1,18 +1,14 @@
-export {
-  ContentProviderRegistry,
-  splitNamespace,
-  DOCS_PROVIDER,
-} from '../doc-engine/index.ts';
+export { DOCS_PROVIDER } from './providers/docs.provider.ts';
 export type {
-  ContentNode,
-  ContentProvider,
-  ContentResolutionRequest,
-  NotFound,
   ContentMeta,
   HeaderDocDefinition,
   HeaderDocLink,
   HeaderDocSection,
-} from '../doc-engine/index.ts';
+} from './packs/header-docs/header-docs.catalog.ts';
+export { HEADER_DOC_DEFINITIONS, resolveHeaderDoc } from './packs/header-docs/header-docs.catalog.ts';
+
+export { ContentProviderRegistry, splitNamespace } from '../doc-engine/index.ts';
+export type { ContentNode, ContentProvider, ContentResolutionRequest, NotFound } from '../doc-engine/index.ts';
 
 // NOTE: contentProviderRegistry is intentionally not re-exported from this barrel.
 // Import it from ./contentEngine to keep a single canonical ownership path.
